🛠️: refactor `DatasetError` test to `typeScript`
